Add admind and admin CLI   * add /v1/admin/account POST
Creates a way for an initial account to be added (and others) with just oauth2 based credentials.
Initial store configuration can therefore be setup without manual intervention.

CLI login will open a browser, send you to the IDP/Keycloak login, redirect you back and store the
credentials in a file. It will then subsequently refresh your token if it has expired as long as
your refresh token has not also expired.

On the Admind side it will verify that your access token is valid before performing an operation.